+++ packages/berlin/src/components/form-input/FormNumberInput.tsx
@@ -1,13 +1,12 @@
 import { Controller, FieldValues, Path, UseFormReturn } from 'react-hook-form';
 import Input from '../input';
 
-export function TextInput<T extends FieldValues>(props: {
+export function FormNumberInput<T extends FieldValues>(props: {
   form: UseFormReturn<T>;
   name: Path<T>;
   label: string;
   required: boolean | null;
-  placeholder?: string;
-  customValidation?: (value: string) => string | undefined;
+  customValidation?: (value: number) => string | undefined;
 }) {
   return (
     <Controller
@@ -19,18 +18,18 @@ export function TextInput<T extends FieldValues>(props: {
       }}
       render={({ field }) => (
         <Input
-          type="text"
+          type="number"
           label={props.label}
           required={!!props.required}
-          placeholder={props.placeholder ?? 'Please enter a value'}
+          placeholder="Enter a value"
           errors={
             props.form.formState.errors[props.name]
               ? [props.form.formState.errors[props.name]?.message?.toString() ?? '']
               : []
           }
           value={field.value ?? undefined}
-          onBlur={field.onBlur}
           onChange={field.onChange}
+          onBlur={field.onBlur}
         />
       )}
     />
