+++ packages/berlin/src/components/form-input/FormTextAreaInput.tsx
@@ -1,32 +1,34 @@
 import { Controller, FieldValues, Path, UseFormReturn } from 'react-hook-form';
-import Select from '../select';
+import Textarea from '../textarea';
 
-export function SelectInput<T extends FieldValues>(props: {
+export function FormTextAreaInput<T extends FieldValues>(props: {
   form: UseFormReturn<T>;
   name: Path<T>;
   label: string;
   required: boolean | null;
-  options: { name: string; value: string }[];
+  customValidation?: (value: string) => string | undefined;
 }) {
   return (
     <Controller
       name={props.name}
       control={props.form.control}
-      rules={{ required: props.required ? 'Value is required' : false }}
+      rules={{
+        required: props.required ? `${props.label} is required` : false,
+        validate: props.customValidation,
+      }}
       render={({ field }) => (
-        <Select
+        <Textarea
           label={props.label}
-          placeholder="Choose a value"
           required={!!props.required}
-          options={props.options.map((option) => ({ id: option.value, name: option.name }))}
+          placeholder="Enter a value"
           errors={
-            props.form.formState.errors[props.name]?.message
+            props.form.formState.errors[props.name]
               ? [props.form.formState.errors[props.name]?.message?.toString() ?? '']
-              : undefined
+              : []
           }
-          onBlur={field.onBlur}
-          onChange={(val) => field.onChange(val)}
           value={field.value ?? undefined}
+          onChange={field.onChange}
+          onBlur={field.onBlur}
         />
       )}
     />
