+++ packages/berlin/src/components/form-input/index.ts
@@ -1,0 +1,5 @@
+export * from './FormNumberInput';
+export * from './FormSelectInput';
+export * from './FormTextAreaInput';
+export * from './FormTextInput';
+export * from './FormInput';
