+++ packages/berlin/src/components/form/index.ts
@@ -1,5 +1,2 @@
-export * from './NumberInput';
-export * from './SelectInput';
-export * from './TextAreaInput';
-export * from './TextInput';
-export * from './FormInput';
+export * from './AccountForm';
+export * from './ResearchGroupForm';
