--- packages/berlin/src/components/research-group-form/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ResearchGroupForm';
